+++ packages/cli/CHANGELOG.md
@@ -17,9 +17,16 @@
 - Added `git:` and `npm:` locators to the `generator.templates` setting in `.moon/workspace.yml`.
   - This allows templates to be packaged and managed outside of the workspace.
   - Locations will be cloned/downloaded on-demand.
+- Added an `id` setting to `template.yml`, so that templates can customize their name (instead of
+  using the folder name).
+- Updated template enum variable `default` fields to support an array of values.
 - Updated task inheritance to support stack-based configuration, such as
   `.moon/tasks/node-frontend.yml` or `.moon/tasks/bun-backend-application.yml`.
 - Updated `moon project` and `moon task` to include the configuration files that tasks inherit from.
+
+#### 🐞 Fixes
+
+- Fixed template enum variable default values being able to use a non-supported value.
 
 #### ⚙️ Internal